(tradeNum, userBuyCost) int에서 long으로 타입 변경, CashDTO long marging(증거금) 추가

+++ src/main/java/com/frade/dto/order/HistoryDTO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class HistoryDTO {	//체결 내역
 	
-	int tradeNum; //거래 번호
+	long tradeNum; //거래 번호
 	int userNum; //유저 번호
 	String stockCode; //종목 코드
 	int tradePrice; //체결 가격
