+++ src/main/java/com/frade/dto/user/PortfolioDTO.java
@@ -8,6 +8,6 @@ public class PortfolioDTO {
 	int userNum; //유저 번호
 	String stockCode; //종목 번호
 	int userStockCnt; //보유 수량
-	int userBuyCost; //현재까지 매수한 가격의 총합
+	long userBuyCost; //현재까지 매수한 가격의 총합
 	
 }
